Permission checks

+++ GearzBukkitPermissions/src/main/java/net/cogz/permissions/bukkit/PermissionsCommands.java
@@ -27,12 +27,14 @@ public class PermissionsCommands implements TCommandHandler {
         switch (args[1]) {
             case "reset":
             case "delete":
+                if (!sender.hasPermission("gearz.permissions.player.delete")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 2) return TCommandStatus.INVALID_ARGS;
                 String name = player.getName();
                 player.remove();
                 permsManager.onJoin(name);
                 sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.player-deleted"));
             case "set":
+                if (!sender.hasPermission("gearz.permissions.player.set")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 3 || args.length > 4) return TCommandStatus.INVALID_ARGS;
                 if (player == null) {
                     sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.null-player", false));
@@ -47,6 +49,8 @@ public class PermissionsCommands implements TCommandHandler {
                 player.addPermission(permission, value);
                 break;
             case "remove":
+            case "unset":
+                if (!sender.hasPermission("gearz.permissions.player.remove")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 if (player == null) {
                     sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.null-player", false));
@@ -56,6 +60,7 @@ public class PermissionsCommands implements TCommandHandler {
                 player.removePermission(args[2]);
                 return TCommandStatus.SUCCESSFUL;
             case "check":
+                if (!sender.hasPermission("gearz.permissions.player.check")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 if (player == null) {
                     sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.null-player", false));
@@ -70,6 +75,7 @@ public class PermissionsCommands implements TCommandHandler {
             case "perms":
             case "show":
             case "permissions":
+                if (!sender.hasPermission("gearz.permissions.player.show")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 2) return TCommandStatus.INVALID_ARGS;
                 for (String perm : player.getPermissions()) {
                     String[] s = perm.split(",");
@@ -78,6 +84,7 @@ public class PermissionsCommands implements TCommandHandler {
                 return TCommandStatus.SUCCESSFUL;
             case "addgroup":
             case "setgroup":
+                if (!sender.hasPermission("gearz.permissions.player.addgroup")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 PermGroup group = permsManager.getGroup(args[2]);
                 if (group == null) {
@@ -88,6 +95,7 @@ public class PermissionsCommands implements TCommandHandler {
                 player.addPlayerToGroup(group);
                 return TCommandStatus.SUCCESSFUL;
             case "prefix":
+                if (!sender.hasPermission("gearz.permissions.player.prefix")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 3) return TCommandStatus.INVALID_ARGS;
                 String prefix;
                 prefix = Gearz.getInstance().compile(args, 2, args.length);
@@ -96,6 +104,7 @@ public class PermissionsCommands implements TCommandHandler {
                 player.save();
                 return TCommandStatus.SUCCESSFUL;
             case "suffix":
+                if (!sender.hasPermission("gearz.permissions.player.suffix")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 3) return TCommandStatus.INVALID_ARGS;
                 String suffix;
                 suffix = Gearz.getInstance().compile(args, 2, args.length);
@@ -104,15 +113,17 @@ public class PermissionsCommands implements TCommandHandler {
                 player.save();
                 return TCommandStatus.SUCCESSFUL;
             case "tabcolor":
+                if (!sender.hasPermission("gearz.permissions.player.tabcolor")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 player.tabColor = args[2];
                 sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.set-tab-color", true, new String[]{"<color>", args[2]}));
                 player.save();
                 return TCommandStatus.SUCCESSFUL;
             case "namecolor":
+                if (!sender.hasPermission("gearz.permissions.player.namecolor")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 player.nameColor = args[2];
-                sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.set-tab-prefix", true, new String[]{"<color>", args[2]}));
+                sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.set-name-color", true, new String[]{"<color>", args[2]}));
                 player.save();
                 return TCommandStatus.SUCCESSFUL;
             default:
@@ -135,6 +146,7 @@ public class PermissionsCommands implements TCommandHandler {
         PermGroup group = permsManager.getGroup(args[0]);
         switch (args[1]) {
             case "create":
+                if (!sender.hasPermission("gearz.permissions.group.create")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 2 || args.length > 3) return TCommandStatus.INVALID_ARGS;
                 boolean defau = false;
                 if (args.length == 3) {
@@ -144,11 +156,13 @@ public class PermissionsCommands implements TCommandHandler {
                 permsManager.createGroup(args[0], defau);
                 return TCommandStatus.SUCCESSFUL;
             case "delete":
+                if (!sender.hasPermission("gearz.permissions.group.delete")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 2) return TCommandStatus.INVALID_ARGS;
                 sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.deleted-group", false, new String[]{"<group>", args[0]}));
                 permsManager.deleteGroup(args[0]);
                 return TCommandStatus.SUCCESSFUL;
             case "set":
+                if (!sender.hasPermission("gearz.permissions.group.set")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 3 || args.length > 4) return TCommandStatus.INVALID_ARGS;
                 if (group == null) {
                     sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.null-group", false));
@@ -163,6 +177,8 @@ public class PermissionsCommands implements TCommandHandler {
                 group.addPermission(permission, value);
                 break;
             case "remove":
+            case "unset":
+                if (!sender.hasPermission("gearz.permissions.group.remove")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 if (group == null) {
                     sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.null-group", false));
@@ -172,6 +188,7 @@ public class PermissionsCommands implements TCommandHandler {
                 group.removePermission(args[2]);
                 return TCommandStatus.SUCCESSFUL;
             case "check":
+                if (!sender.hasPermission("gearz.permissions.group.check")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 if (group == null) {
                     sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.null-group", false));
@@ -186,6 +203,7 @@ public class PermissionsCommands implements TCommandHandler {
             case "perms":
             case "show":
             case "permissions":
+                if (!sender.hasPermission("gearz.permissions.group.show")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 2) return TCommandStatus.INVALID_ARGS;
                 for (String perm : group.getPermissions()) {
                     String[] s = perm.split(",");
@@ -193,6 +211,7 @@ public class PermissionsCommands implements TCommandHandler {
                 }
                 return TCommandStatus.SUCCESSFUL;
             case "prefix":
+                if (!sender.hasPermission("gearz.permissions.group.prefix")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 3) return TCommandStatus.INVALID_ARGS;
                 String prefix;
                 prefix = Gearz.getInstance().compile(args, 2, args.length);
@@ -201,6 +220,7 @@ public class PermissionsCommands implements TCommandHandler {
                 group.save();
                 return TCommandStatus.SUCCESSFUL;
             case "suffix":
+                if (!sender.hasPermission("gearz.permissions.group.suffix")) return TCommandStatus.PERMISSIONS;
                 if (args.length < 3) return TCommandStatus.INVALID_ARGS;
                 String suffix;
                 suffix = Gearz.getInstance().compile(args, 2, args.length);
@@ -209,15 +229,17 @@ public class PermissionsCommands implements TCommandHandler {
                 group.save();
                 return TCommandStatus.SUCCESSFUL;
             case "tabcolor":
+                if (!sender.hasPermission("gearz.permissions.group.tabcolor")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 group.tabColor = args[2];
                 sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.set-tab-color", true, new String[]{"<color>", args[2]}));
                 group.save();
                 return TCommandStatus.SUCCESSFUL;
             case "namecolor":
+                if (!sender.hasPermission("gearz.permissions.group.namecolor")) return TCommandStatus.PERMISSIONS;
                 if (args.length != 3) return TCommandStatus.INVALID_ARGS;
                 group.nameColor = args[2];
-                sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.set-tab-prefix", true, new String[]{"<color>", args[2]}));
+                sender.sendMessage(GearzBukkitPermissions.getInstance().getFormat("formats.set-name-color", true, new String[]{"<color>", args[2]}));
                 group.save();
                 return TCommandStatus.SUCCESSFUL;
             default:
@@ -248,6 +270,4 @@ public class PermissionsCommands implements TCommandHandler {
     public void handleCommandStatus(TCommandStatus status, CommandSender sender, TCommandSender senderType) {
         Gearz.getInstance().handleCommandStatus(status, sender, senderType);
     }
-
-
 }
